fully port segment display and acceleration logic

+++ src/hooks/acceleration.rs
@@ -2,9 +2,9 @@ use chrono::Duration;
 use leptos::prelude::*;
 
 pub fn use_acceleration(
-  base_increment: f64,
+  base_increment: f32,
   timeout_ms: Duration,
-) -> impl Fn() -> f64 {
+) -> impl Fn() -> f32 {
   let (last_time, set_last_time) = signal(chrono::Local::now());
 
   move || {
@@ -16,7 +16,7 @@ pub fn use_acceleration(
     if diff > timeout_ms || diff < Duration::milliseconds(1) {
       base_increment
     } else {
-      (timeout_ms.num_milliseconds() as f64 / diff.num_milliseconds() as f64)
+      (timeout_ms.num_milliseconds() as f32 / diff.num_milliseconds() as f32)
         .powf(2.0)
         * base_increment
     }
